penalise missed boreholes as false negatives
